ads/views.py
from rest_framework import viewsets
from .models import Location,Ad
from .serializers import AdSerializer, LocationSerializer

class AdViewSet(viewsets.ModelViewSet):
    queryset = Ad.objects.all()
    serializer_class = AdSerializer
    
class LocationViewSet(viewsets.ModelViewSet):
    queryset = Location.objects.all()
    serializer_class = LocationSerializer 

